Delete %%3F^3F8^3F8E1A98%%playlist_item.html.php
Removed files in templates_c directory. If these files are present any edited files must be updated with the 'touch' command before changes will be recognized. This is because the files in the templates_c directory have the same modify date and time and therefore will not be rebuilt to accommodate any changes made in the editing process.